+++ src/morelia/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from .base import *  # noqa
+from .parser import Parser  # noqa